// frontend/src/services/typescript-axios/models/get-ward-patient.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface GetWardPatient
 */
export interface GetWardPatient {
    /**
     * 
     * @type {string}
     * @memberof GetWardPatient
     */
    patientKey?: string;
    /**
     * 
     * @type {Date}
     * @memberof GetWardPatient
     */
    admissionTime?: Date;
    /**
     * 
     * @type {string}
     * @memberof GetWardPatient
     */
    bedNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof GetWardPatient
     */
    caseNumber?: string;
    /**
     * 
     * @type {Date}
     * @memberof GetWardPatient
     */
    dob: Date;
    /**
     * 
     * @type {string}
     * @memberof GetWardPatient
     */
    patientName?: string;
    /**
     * 
     * @type {string}
     * @memberof GetWardPatient
     */
    sex?: string;
    /**
     * 
     * @type {string}
     * @memberof GetWardPatient
     */
    specialtyCode?: string;
    /**
     * 
     * @type {string}
     * @memberof GetWardPatient
     */
    wardCode?: string;
    /**
     * 
     * @type {Date}
     * @memberof GetWardPatient
     */
    updateDate?: Date;
}